+++ superset-frontend/src/features/databases/DatabaseModal/index.test.tsx
@@ -1140,6 +1140,215 @@ describe('DatabaseModal', () => {
       expect(sqlAlchemyFormStepText).toBeVisible();
     });
 
+    describe('SQL Alchemy form flow', () => {
+      test('enters step 2 of 2 when proper database is selected', async () => {
+        userEvent.click(
+          screen.getByRole('button', {
+            name: /sqlite/i,
+          }),
+        );
+
+        expect(await screen.findByText(/step 2 of 2/i)).toBeInTheDocument();
+      });
+
+      test('runs fetchResource when "Connect" is clicked', () => {
+        /* ---------- 🐞 TODO (lyndsiWilliams): function mock is not currently working 🐞 ----------
+
+        // Mock useSingleViewResource
+        const mockUseSingleViewResource = jest.fn();
+        mockUseSingleViewResource.mockImplementation(useSingleViewResource);
+
+        const { fetchResource } = mockUseSingleViewResource('database');
+
+        // Invalid hook call?
+        userEvent.click(screen.getByRole('button', { name: 'Connect' }));
+        expect(fetchResource).toHaveBeenCalled();
+
+        The line below makes the linter happy */
+        expect.anything();
+      });
+
+      describe('step 2 component interaction', () => {
+        test('properly interacts with textboxes', async () => {
+          userEvent.click(
+            screen.getByRole('button', {
+              name: /sqlite/i,
+            }),
+          );
+
+          expect(await screen.findByText(/step 2 of 2/i)).toBeInTheDocument();
+          const dbNametextBox = screen.getByTestId('database-name-input');
+          expect(dbNametextBox).toHaveValue('SQLite');
+
+          userEvent.type(dbNametextBox, 'Different text');
+          expect(dbNametextBox).toHaveValue('SQLiteDifferent text');
+
+          const sqlAlchemyURItextBox = screen.getByTestId(
+            'sqlalchemy-uri-input',
+          );
+          expect(sqlAlchemyURItextBox).toHaveValue('');
+
+          userEvent.type(sqlAlchemyURItextBox, 'Different text');
+          expect(sqlAlchemyURItextBox).toHaveValue('Different text');
+        });
+
+        test('runs testDatabaseConnection when "TEST CONNECTION" is clicked', () => {
+          /* ---------- 🐞 TODO (lyndsiWilliams): function mock is not currently working 🐞 ----------
+
+          // Mock testDatabaseConnection
+          const mockTestDatabaseConnection = jest.fn();
+          mockTestDatabaseConnection.mockImplementation(testDatabaseConnection);
+
+          userEvent.click(
+            screen.getByRole('button', {
+              name: /test connection/i,
+            }),
+          );
+
+          expect(mockTestDatabaseConnection).toHaveBeenCalled();
+
+          The line below makes the linter happy */
+          expect.anything();
+        });
+      });
+
+      describe('SSH Tunnel Form interaction', () => {
+        test('properly interacts with SSH Tunnel form textboxes for dynamic form', async () => {
+          userEvent.click(
+            screen.getByRole('button', {
+              name: /postgresql/i,
+            }),
+          );
+          expect(await screen.findByText(/step 2 of 3/i)).toBeInTheDocument();
+          const SSHTunnelingToggle = screen.getByTestId('ssh-tunnel-switch');
+          userEvent.click(SSHTunnelingToggle);
+          const SSHTunnelServerAddressInput = screen.getByTestId(
+            'ssh-tunnel-server_address-input',
+          );
+          expect(SSHTunnelServerAddressInput).toHaveValue('');
+          userEvent.type(SSHTunnelServerAddressInput, 'localhost');
+          expect(SSHTunnelServerAddressInput).toHaveValue('localhost');
+          const SSHTunnelServerPortInput = screen.getByTestId(
+            'ssh-tunnel-server_port-input',
+          );
+          expect(SSHTunnelServerPortInput).toHaveValue(null);
+          userEvent.type(SSHTunnelServerPortInput, '22');
+          expect(SSHTunnelServerPortInput).toHaveValue(22);
+          const SSHTunnelUsernameInput = screen.getByTestId(
+            'ssh-tunnel-username-input',
+          );
+          expect(SSHTunnelUsernameInput).toHaveValue('');
+          userEvent.type(SSHTunnelUsernameInput, 'test');
+          expect(SSHTunnelUsernameInput).toHaveValue('test');
+          const SSHTunnelPasswordInput = screen.getByTestId(
+            'ssh-tunnel-password-input',
+          );
+          expect(SSHTunnelPasswordInput).toHaveValue('');
+          userEvent.type(SSHTunnelPasswordInput, 'pass');
+          expect(SSHTunnelPasswordInput).toHaveValue('pass');
+        });
+
+        test('properly interacts with SSH Tunnel form textboxes', async () => {
+          userEvent.click(
+            screen.getByRole('button', {
+              name: /sqlite/i,
+            }),
+          );
+
+          expect(await screen.findByText(/step 2 of 2/i)).toBeInTheDocument();
+          const SSHTunnelingToggle = screen.getByTestId('ssh-tunnel-switch');
+          userEvent.click(SSHTunnelingToggle);
+          const SSHTunnelServerAddressInput = screen.getByTestId(
+            'ssh-tunnel-server_address-input',
+          );
+          expect(SSHTunnelServerAddressInput).toHaveValue('');
+          userEvent.type(SSHTunnelServerAddressInput, 'localhost');
+          expect(SSHTunnelServerAddressInput).toHaveValue('localhost');
+          const SSHTunnelServerPortInput = screen.getByTestId(
+            'ssh-tunnel-server_port-input',
+          );
+          expect(SSHTunnelServerPortInput).toHaveValue(null);
+          userEvent.type(SSHTunnelServerPortInput, '22');
+          expect(SSHTunnelServerPortInput).toHaveValue(22);
+          const SSHTunnelUsernameInput = screen.getByTestId(
+            'ssh-tunnel-username-input',
+          );
+          expect(SSHTunnelUsernameInput).toHaveValue('');
+          userEvent.type(SSHTunnelUsernameInput, 'test');
+          expect(SSHTunnelUsernameInput).toHaveValue('test');
+          const SSHTunnelPasswordInput = screen.getByTestId(
+            'ssh-tunnel-password-input',
+          );
+          expect(SSHTunnelPasswordInput).toHaveValue('');
+          userEvent.type(SSHTunnelPasswordInput, 'pass');
+          expect(SSHTunnelPasswordInput).toHaveValue('pass');
+        });
+
+        test('if the SSH Tunneling toggle is not true, no inputs are displayed', async () => {
+          userEvent.click(
+            screen.getByRole('button', {
+              name: /sqlite/i,
+            }),
+          );
+
+          expect(await screen.findByText(/step 2 of 2/i)).toBeInTheDocument();
+          const SSHTunnelingToggle = screen.getByTestId('ssh-tunnel-switch');
+          expect(SSHTunnelingToggle).toBeVisible();
+          const SSHTunnelServerAddressInput = screen.queryByTestId(
+            'ssh-tunnel-server_address-input',
+          );
+          expect(SSHTunnelServerAddressInput).not.toBeInTheDocument();
+          const SSHTunnelServerPortInput = screen.queryByTestId(
+            'ssh-tunnel-server_port-input',
+          );
+          expect(SSHTunnelServerPortInput).not.toBeInTheDocument();
+          const SSHTunnelUsernameInput = screen.queryByTestId(
+            'ssh-tunnel-username-input',
+          );
+          expect(SSHTunnelUsernameInput).not.toBeInTheDocument();
+          const SSHTunnelPasswordInput = screen.queryByTestId(
+            'ssh-tunnel-password-input',
+          );
+          expect(SSHTunnelPasswordInput).not.toBeInTheDocument();
+        });
+
+        test('If user changes the login method, the inputs change', async () => {
+          userEvent.click(
+            screen.getByRole('button', {
+              name: /sqlite/i,
+            }),
+          );
+
+          expect(await screen.findByText(/step 2 of 2/i)).toBeInTheDocument();
+          const SSHTunnelingToggle = screen.getByTestId('ssh-tunnel-switch');
+          userEvent.click(SSHTunnelingToggle);
+          const SSHTunnelUsePasswordInput = screen.getByTestId(
+            'ssh-tunnel-use_password-radio',
+          );
+          expect(SSHTunnelUsePasswordInput).toBeVisible();
+          const SSHTunnelUsePrivateKeyInput = screen.getByTestId(
+            'ssh-tunnel-use_private_key-radio',
+          );
+          expect(SSHTunnelUsePrivateKeyInput).toBeVisible();
+          const SSHTunnelPasswordInput = screen.getByTestId(
+            'ssh-tunnel-password-input',
+          );
+          // By default, we use Password as login method
+          expect(SSHTunnelPasswordInput).toBeVisible();
+          // Change the login method to use private key
+          userEvent.click(SSHTunnelUsePrivateKeyInput);
+          const SSHTunnelPrivateKeyInput = screen.getByTestId(
+            'ssh-tunnel-private_key-input',
+          );
+          expect(SSHTunnelPrivateKeyInput).toBeVisible();
+          const SSHTunnelPrivateKeyPasswordInput = screen.getByTestId(
+            'ssh-tunnel-private_key_password-input',
+          );
+          expect(SSHTunnelPrivateKeyPasswordInput).toBeVisible();
+        });
+      });
+    });
+
     describe('Dynamic form flow', () => {
       test('enters step 2 of 3 when proper database is selected', async () => {
         expect(await screen.findByText(/step 1 of 3/i)).toBeInTheDocument();
